+++ Pixl/Sources/Pixl2D/Shapes/CutDisk.swift
@@ -1,16 +1,16 @@
 /// Disk cut by a horizontal chord.
 public struct CutDisk: Hashable, Sendable {
     /// Disk radius.
-    public let radius: Double
+    public let radius: Float
     /// Signed chord height from the centre.
-    public let height: Double
+    public let height: Float
     /// Creates a canonical unit half disk.
     public init() { self.init(radius: 0.5, height: 0) }
     /// Creates a cut disk.
     /// - Parameters:
     ///   - radius: Positive disk radius.
     ///   - height: Finite chord height with magnitude no greater than `radius`.
-    public init(radius: Double, height: Double) {
+    public init(radius: Float, height: Float) {
         precondition(radius.isFinite && radius > 0 && height.isFinite && abs(height) <= radius)
         self.radius = radius; self.height = height
     }
@@ -20,5 +20,5 @@ public struct CutDisk: Hashable, Sendable {
     /// - Parameters:
     ///   - radius: Positive disk radius.
     ///   - height: Finite chord height with magnitude no greater than `radius`.
-    public static func cutDisk(radius: Double, height: Double) -> Self { .init(radius: radius, height: height) }
+    public static func cutDisk(radius: Float, height: Float) -> Self { .init(radius: radius, height: height) }
 }
