+++ Pixl/Sources/Pixl2D/Shapes/Horseshoe.swift
@@ -3,13 +3,13 @@ import PixlGraphics
 /// Horseshoe geometry cut from a thick circular arc.
 public struct Horseshoe: Hashable, Sendable {
     /// Centre-line radius.
-    public let radius: Double
+    public let radius: Float
     /// Opening orientation.
     public let angle: Angle
     /// Straight-end length.
-    public let length: Double
+    public let length: Float
     /// Full band width.
-    public let width: Double
+    public let width: Float
     /// Creates a canonical unit horseshoe.
     public init() { self.init(radius: 0.35, angle: .degrees(120), length: 0.15, width: 0.1) }
     /// Creates a horseshoe.
@@ -18,7 +18,7 @@ public struct Horseshoe: Hashable, Sendable {
     ///   - angle: Finite opening orientation.
     ///   - length: Nonnegative straight-end length.
     ///   - width: Positive full line width.
-    public init(radius: Double, angle: Angle, length: Double, width: Double) {
+    public init(radius: Float, angle: Angle, length: Float, width: Float) {
         precondition(radius.isFinite && radius > 0 && angle.radians.isFinite)
         precondition(length.isFinite && length >= 0 && width.isFinite && width > 0)
         self.radius = radius; self.angle = angle; self.length = length; self.width = width
@@ -31,7 +31,7 @@ public struct Horseshoe: Hashable, Sendable {
     ///   - angle: Finite opening orientation.
     ///   - length: Nonnegative straight-end length.
     ///   - width: Positive full band width.
-    public static func horseshoe(radius: Double, angle: Angle, length: Double, width: Double) -> Self {
+    public static func horseshoe(radius: Float, angle: Angle, length: Float, width: Float) -> Self {
         .init(radius: radius, angle: angle, length: length, width: width)
     }
 }
